Modified FractureSet.Calculate_MeanPropagationDistance() so that when it checks the mean propagation rate is greater than zero, it converts to a float to eliminate rounding errors; this prevents it from returning infinite values (B88). Also modified FractureDipSet.calculateTotalMacrofracturePopulation() and FractureDipSet.calculateCumulativeMacrofracturePopulationArrays() to calculate zero values and increments for P30 and P32 fracture densities if the fracture sets are not activated, and modified FractureDipSet.setMacrofractureDeactivationRate() to more efficiently calculate the mean active fracture half length by not looking up the MFP30 density twice.

+++ DFMGenerator_SharedCode/FractureSet.cs
@@ -2837,7 +2837,7 @@ namespace DFMGenerator_SharedCode
             // Get the mean propagation rate
             // NB if this is zero the mean propagation distances will all be zero so we can create an array of zero values and return this
             double propagationRate = (UseCurrentData ? mp_fds.getMeanMFPropagationRate() : mp_fds.getMeanMFPropagationRate(Timestep_M));
-            if (propagationRate <= 0)
+            if ((float)propagationRate <= 0f)
             {
                 foreach (double cutoff in cutoffLengths)
                     meanPropagationDistances.Add(0);
